+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-
+### Added
+- `Oauth2ServerFactory::create()` now accepts optional `$customGrants` parameter for injecting custom grant types
 
 ## [1.0.0] - 2026-01-24